+++ tests/data/drivers/test_kql_driver.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-"""datq query test class."""
+"""KQL driver query test class."""
 from contextlib import redirect_stdout
 import io
 from unittest.mock import patch
@@ -127,8 +127,46 @@ def kql_exec(content):
 KQL_EXEC_PATCH = (kql_driver, "kql_exec", kql_exec)
 
 
+class AzCredentials:
+    """Mock credentials class."""
+
+    class ModernCred:
+        """Mock modern credentials class."""
+
+        class Token:
+            """Mocked token class."""
+
+            token = "Token"
+
+        @classmethod
+        def get_token(cls, *args, **kwargs):
+            """Return the token."""
+            del args, kwargs
+            return cls.Token()
+
+        @property
+        def credentials(self):
+            """Return mocked credentials list"""
+            return ["cred1", "cred2", "cred3"]
+
+    @property
+    def modern(self):
+        """Return the modern credentials."""
+        return self.ModernCred()
+
+
+def az_connect(*args, **kwargs):
+    """Mock the az_connect function."""
+    del args, kwargs
+    return AzCredentials()
+
+
+AZ_CONNECT_PATH = (kql_driver, "az_connect", az_connect)
+
+
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_load(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -143,6 +181,7 @@ def test_kql_load(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_connect(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -155,6 +194,7 @@ def test_kql_connect(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_connect_no_cs(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -167,6 +207,7 @@ def test_kql_connect_no_cs(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_connect_kql_exceptions(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -194,6 +235,7 @@ def test_kql_connect_kql_exceptions(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_connect_adal_exceptions(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -218,6 +260,7 @@ def test_kql_connect_adal_exceptions(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_connect_authn_exceptions(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -231,6 +274,7 @@ def test_kql_connect_authn_exceptions(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_schema(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -244,6 +288,7 @@ def test_kql_schema(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_query_not_connected(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -257,6 +302,7 @@ def test_kql_query_not_connected(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_query_failed(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -276,6 +322,7 @@ def test_kql_query_failed(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_query_success(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -288,6 +335,7 @@ def test_kql_query_success(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_query_partial(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
@@ -303,6 +351,7 @@ def test_kql_query_partial(get_ipython):
 
 @patch(GET_IPYTHON_PATCH)
 @patch.object(*KQL_EXEC_PATCH)
+@patch.object(*AZ_CONNECT_PATH)
 def test_kql_query_no_table(get_ipython):
     """Check loaded true."""
     get_ipython.return_value = _MockIPython()
